+++ CHANGELOG.md
@@ -15,11 +15,10 @@
 - adds a `dataclass_serialization_class` utility function that automatically generates the asdf serialization class for 
   python dataclasses. [[#380]](https://github.com/BAMWelDX/weldx/pull/380)
 - Added method to set the interpolation method to the `TimeSeries` [[#353]](https://github.com/BAMWelDX/weldx/pull/353)
+- Add `is_discrete` and `is_expression` properties to `TimeSeries` [[#366]](https://github.com/BAMWelDX/weldx/pull/366)
 
 ### changes
 
-- `LocalCoordinateSystem.__init__` now accepts a `TimeSeries` as input. 
-  [[#366]](https://github.com/BAMWelDX/weldx/pull/366)
 - `WXRotation.from_euler()` now accepts a `pint.Quantity` as input. [[#318]](https://github.com/BAMWelDX/weldx/pull/318)
 - move tests folder to `weldx/tests` [[#323]](https://github.com/BAMWelDX/weldx/pull/323)
 - `get_yaml_header` received a new option parse, which optionally returns the parsed YAML header
@@ -37,7 +36,9 @@
     - Added additional functionality to `Signal`, `SignalTransformation` and `GenericEquipment`
     - Removed `Data` class
     - Updated asdf schemas of all modified classes and the ones that contained references to those classes
-
+- `LocalCoordinateSystem.__init__` now accepts a `TimeSeries` as input. All methods of the `CoordinateSystemManager` 
+  also support this new behavior [[#366]](https://github.com/BAMWelDX/weldx/pull/366)
+  
 ### documentation
 
 - Add new tutorial about the `MeasurementChain` [[#326]](https://github.com/BAMWelDX/weldx/pull/326)